// Day22/challenges_info.d.ts
export declare const asabenehChallenges2020: {
    description: string;
    challengeTitle: string;
    challengeSubtitle: string;
    challengeYear: number;
    keywords: string[];
    author: {
        firstName: string;
        lastName: string;
        titles: string[][];
        qualifications: string[];
        socialLinks: {
            social: string;
            url: string;
            fontawesomeIcon: string;
        }[];
        skills: string[];
        bio: string;
    };
    challenges: {
        name: string;
        topics: string[];
        days: number;
        status: string;
        questions: string;
        projects: string;
        interviewQns: string;
        githubUrl: string;
    }[];
};
//# sourceMappingURL=challenges_info.d.ts.map